--- Models/recipeModel
@@ -1,40 +1,0 @@
-const DBConnections = require('./db');
-const sql = DBConnections.mySqlConnection;
-const mongo = DBConnections.mongoConnection;
-
-
-const details = function (details) {
-    //to add more in Jing Fang portion
-    this.details = details;
-};
-
-details.details = (body) => {
-    return new Promise((resolve, reject) => {
-        mongo.collection("Recipe_Nutrition_Ingredient_Duration").aggregate([{
-
-            $lookup: {
-                from: "Recipe_Review",
-                localField: "recipeID",
-                foreignField: "recipeID",
-                as: "RecipeDetails"
-            }
-
-        }, {
-            $match: {
-                recipeID: parseInt(body.recipeID)
-            }
-        }
-        ]).toArray((err, result) => {
-            if (err) {
-                reject(err)
-            }
-            else {
-                resolve(result)
-            }
-        })
-    })
-}
-
-
-
-module.exports = details;
