+++ Views/assets/css/style.css
@@ -5341,7 +5341,7 @@ td.remove{
 }
 .rate:not(:checked) > input {
   position:absolute;
-  top:-9999px;
+  /* top:-9999px; */
 }
 .rate:not(:checked) > label {
   float:right;
