+++ routes/recipes.js
@@ -14,4 +14,8 @@ router.post("/details",recipe.details);
 router.get("/search-result",recipe.search);
 
 
+router.post("/",recipe.comment_post);
+router.post("/update",recipe.comment_update);
+router.post("/delete",recipe.comment_delete);
+
 module.exports = router;
